requirements.txt
# Python 3.6.9
numpy==1.18.1
pypng==0.0.20
